:sparkles: Update test

+++ tests/Install/InstallTest.php
@@ -10,6 +10,7 @@
 
 namespace Juzaweb\Tests\Install;
 
+use Illuminate\Support\Facades\DB;
 use Juzaweb\Support\Installer;
 use Juzaweb\Tests\TestCase;
 
@@ -17,24 +18,30 @@ class InstallTest extends TestCase
 {
     public function testInstallCommand()
     {
+        $this->resetTestData();
+
         $this->artisan('juzaweb:install')
             ->expectsQuestion('Full Name?', 'Taylor Otwell')
             ->expectsQuestion('Email?', 'demo@gmail.com')
             ->expectsQuestion('Password?', '12345678')
             ->assertExitCode(0);
+
+        $this->assertTrue(file_exists(Installer::installedPath()));
     }
 
-    /*public function testInstallWeb()
+    public function testInstallWeb()
     {
+        $this->resetTestData();
+
         $response = $this->post('install/environment', [
             '_token' => csrf_token(),
-            'database_connection' => 'mysql',
-            'database_hostname' => 'localhost',
-            'database_port' => '3308',
-            'database_name' => 'testcms',
-            'database_username' => 'root',
-            'database_password' => '',
-            'database_prefix' => 'jww_',
+            'database_connection' => env('DB_CONNECTION'),
+            'database_hostname' => env('DB_HOST'),
+            'database_port' => env('DB_PORT'),
+            'database_name' => env('DB_DATABASE'),
+            'database_username' => env('DB_USERNAME'),
+            'database_password' => env('DB_PASSWORD'),
+            'database_prefix' => env('DB_PREFIX'),
         ]);
 
         $response->assertSessionDoesntHaveErrors();
@@ -61,13 +68,17 @@ class InstallTest extends TestCase
         $this->assertTrue(DB::table('users')
             ->where('is_admin', '=', 1)
             ->exists());
-    }*/
+
+        $this->assertTrue(file_exists(Installer::installedPath()));
+    }
 
     protected function resetTestData()
     {
         $this->artisan('migrate:reset')
             ->assertExitCode(0);
 
-        unlink(Installer::installedPath());
+        if (file_exists(Installer::installedPath())) {
+            unlink(Installer::installedPath());
+        }
     }
 }
